[fix] hook생성 및 코드 수정
